move repeated where clause code to its own helper file
