+++ server/models/workers.js
@@ -1,4 +1,5 @@
 import { db_connect } from "../db/database.js"
+import { buildWhereClauses } from "../helpers/whereClauseBuilder.js"
 
 export const queryWorker = async (workerIds, locationIds, excludeIncomplete) => {
   const db = await db_connect();
@@ -23,26 +24,4 @@ const buildDbQuery = async (whereClauses) => {
                        GROUP BY w.id`
                       console.log(queryString);
   return queryString;
-};
-
-const buildWhereClauses = async (workerIds, locationIds, excludeIncomplete) => {
-  const whereClauses = [];
-
-  if (typeof workerIds != 'undefined') {
-    whereClauses.push(`w.id IN (${workerIds})`);
-  }; 
-
-  if (typeof locationIds != 'undefined') {
-    whereClauses.push(`lt.location_id IN (${locationIds})`)
-  };
-
-  if (typeof excludeIncomplete != 'undefined' && excludeIncomplete === true) {
-    whereClauses.push("t.completed_at IS NOT NULL");
-  };
-
-  if (whereClauses.length >= 1) {
-    return (`WHERE ${whereClauses.join(' AND ')}`)
-  } else {
-    return ""
-  };
 };
